fix: change type of nyrs from double to integer

+++ R/fimsframe.R
@@ -12,9 +12,11 @@
 
 setClass(
   Class = "FIMSFrame",
-  slots = c(data = "data.frame", # can use c( ) or list here.
-            fleets = "numeric",
-            nyrs = "numeric")
+  slots = c(
+    data = "data.frame", # can use c( ) or list here.
+    fleets = "numeric",
+    nyrs = "integer"
+  )
 )
 
 # leaving FIMSFrameAge with just age related slots.
@@ -64,20 +66,23 @@ setGeneric("weightatage", function(x) standardGeneric("weightatage"))
 setMethod("weightatage", "FIMSFrameAge", function(x) x@weightatage)
 
 setGeneric("m_weightatage", function(x) standardGeneric("m_weightatage"))
-setMethod("m_weightatage", "FIMSFrameAge",
+setMethod(
+  "m_weightatage", "FIMSFrameAge",
   function(x) {
     dplyr::filter(
       .data = as.data.frame(x@data),
       type == "weight-at-age"
     ) %>%
-    dplyr::group_by(age) %>%
-    dplyr::summarize(mean_value = mean(value)) %>%
-    dplyr::pull(mean_value)
+      dplyr::group_by(age) %>%
+      dplyr::summarize(mean_value = mean(value)) %>%
+      dplyr::pull(mean_value)
   }
 )
 
 setGeneric("m_ages", function(x) standardGeneric("m_ages"))
-setMethod("m_ages", "FIMSFrameAge", function(x) {x@ages})
+setMethod("m_ages", "FIMSFrameAge", function(x) {
+  x@ages
+})
 
 # Note: don't include setters, because for right now, we don't want users to be
 # setting ages, fleets, etc. However, we could allow it in the future, if there
@@ -231,41 +236,41 @@ setValidity(
 #' called `data` to store the input data frame. Additional slots are dependent
 #' on the child class. Use [showClass()] to see all available slots.
 FIMSFrame <- function(data) {
-  #Get the earliest and latest year of data and use to calculate n years for population simulation
-  start_yr <- as.numeric(strsplit(min(data[["datestart"]],na.rm=TRUE),"-")[[1]][1])
-  end_yr <- as.numeric(strsplit(max(data[["dateend"]],na.rm=TRUE),"-")[[1]][1])
-  nyrs <- as.integer(end_yr-start_yr+1)
+  # Get the earliest and latest year of data and use to calculate n years for population simulation
+  start_yr <- as.numeric(strsplit(min(data[["datestart"]], na.rm = TRUE), "-")[[1]][1])
+  end_yr <- as.numeric(strsplit(max(data[["dateend"]], na.rm = TRUE), "-")[[1]][1])
+  nyrs <- as.integer(end_yr - start_yr + 1)
   years <- start_yr:end_yr
 
-  #Get the fleets represented in the data
-  fleets <- unique(data[["name"]])[grep("fleet",unique(data[["name"]]))]
-  fleets <- as.numeric(unlist(lapply(strsplit(fleets,"fleet"),function(x)x[2])))
+  # Get the fleets represented in the data
+  fleets <- unique(data[["name"]])[grep("fleet", unique(data[["name"]]))]
+  fleets <- as.numeric(unlist(lapply(strsplit(fleets, "fleet"), function(x) x[2])))
   nfleets <- length(fleets)
-  #Make empty NA data frames in the format needed to pass to FIMS
+  # Make empty NA data frames in the format needed to pass to FIMS
 
-  #Fill the empty data frames with data extracted from the data file
+  # Fill the empty data frames with data extracted from the data file
   out <- new("FIMSFrame",
-             data = data,
-             fleets = fleets,
-             nyrs = nyrs
-             )
+    data = data,
+    fleets = fleets,
+    nyrs = nyrs
+  )
   return(out)
 }
 #' FIMSFrameAge
 #' @export
 #' @rdname FIMSFrame
 FIMSFrameAge <- function(data) {
-  #Get the earliest and latest year of data and use to calculate n years for population simulation
-  start_yr <- as.numeric(strsplit(min(data[["datestart"]],na.rm=TRUE),"-")[[1]][1])
-  end_yr <- as.numeric(strsplit(max(data[["dateend"]],na.rm=TRUE),"-")[[1]][1])
-  nyrs <- as.integer(end_yr-start_yr+1)
+  # Get the earliest and latest year of data and use to calculate n years for population simulation
+  start_yr <- as.numeric(strsplit(min(data[["datestart"]], na.rm = TRUE), "-")[[1]][1])
+  end_yr <- as.numeric(strsplit(max(data[["dateend"]], na.rm = TRUE), "-")[[1]][1])
+  nyrs <- as.integer(end_yr - start_yr + 1)
   years <- start_yr:end_yr
-  #Get the fleets represented in the data
-  fleets <- unique(data[["name"]])[grep("fleet",unique(data[["name"]]))]
-  fleets <- as.numeric(unlist(lapply(strsplit(fleets,"fleet"),function(x)x[2])))
+  # Get the fleets represented in the data
+  fleets <- unique(data[["name"]])[grep("fleet", unique(data[["name"]]))]
+  fleets <- as.numeric(unlist(lapply(strsplit(fleets, "fleet"), function(x) x[2])))
   nfleets <- length(fleets)
-  #Make empty NA data frames in the format needed to pass to FIMS
-  #Get the range of ages displayed in the data to use to specify population simulation range
+  # Make empty NA data frames in the format needed to pass to FIMS
+  # Get the range of ages displayed in the data to use to specify population simulation range
   ages <- min(data[["age"]], na.rm = TRUE):max(data[["age"]], na.rm = TRUE)
   nages <- length(ages)
   weightatage <- dplyr::filter(
